perf(parsing): stop rebuilding regexes and prefixes on hot VQL/import paths
REVIEW.md PF-2..PF-5: binary-search a memoized newline-offset array in the
hunt-query tokenizer's location() instead of re-splitting the whole prefix
per token (O(n^2) -> O(n log n), byte-identical positions); memoize the
EVTX elText/attr patterns and count events with an exec loop instead of a
throwaway match() array; precompile the Velociraptor message-field label
regexes at module level; and resolve getFlowInfo's hostname with a targeted
clients(search='id:...') single-row query instead of enumerating up to 100k
clients through a subprocess. veloMessageFields.ts and clientInventory.ts
are verbatim extractions to satisfy the frozen file-size ledger; existing
import paths are preserved via re-exports.
